Split transport wire encryption policy
