+++ crates/libs/rns-transport/src/transport/mod.rs
@@ -311,6 +311,8 @@ mod links;
 mod path;
 // resource_wire: link-scoped resource packet handling on inbound wire paths.
 mod resource_wire;
+// wire_encryption: packet-context policy for transport-layer encryption.
+mod wire_encryption;
 // wire: inbound packet handlers and wire-level packet logic.
 mod wire;
